// front-end/src/app/models/properties.ts
export interface Properties {
    propertyid: any;
    owner: any; // Replace 'any' with the type of 'owner'
    agent: any; // Replace 'any' with the type of 'agent'
    addressline1: any;
    addressline2: any;
    state: any;
    city: any;
    pincode: any;
    size: any;
    price: any;
    propertytype: any;
    propstatus: any;
    images: any;
    ratings: any[]; // Replace 'any' with the type of 'ratings'
    transactions: any[]; // Replace 'any' with the type of 'transactions'
    userHasTransaction:any;
  }